added uploader details and code refactored
