+++ angular-src/src/app/components/upload/upload.component.ts
@@ -7,7 +7,9 @@ import { FileUploader } from 'ng2-file-upload';
   styleUrls: ['./upload.component.css']
 })
 export class UploadComponent implements OnInit {
-  public uploader:FileUploader = new FileUploader({url:'http://localhost:3000/clients/upload'});
+  user: any;
+
+  public uploader:FileUploader = new FileUploader({url:'http://localhost:3000/clients/upload'+JSON.parse(localStorage.getItem('user')).username});
   constructor() { }
 
   ngOnInit() {
